requirements.txt
requests
dnspython
pytest
python-consul
